+++ config/splat.us.stno3.yaml
@@ -51,7 +51,7 @@ segments:
       - [0xEA0, .data, entrance_stage_entities]
       - [0x1044, .data, e_castle_door]
       - [0x1140, .data, e_background_bushes_trees]
-      - [0x123C, .data, entrance_stage_entities_3]
+      - [0x123C, .data, e_trapdoor]
       - [0x1258, .data, e_jewel_sword_puzzle]
       - [0x1358, .data, e_sky_entities]
       - [0x13A8, data]
@@ -170,7 +170,7 @@ segments:
       - [0x39198, c, e_transparent_water]
       - [0x394F0, c, e_cavern_door]
       - [0x39C44, c, entrance_weights]
-      - [0x3A348, c, entrance_stage_entities_3]
+      - [0x3A348, c, e_trapdoor]
       - [0x3A4C0, c, e_jewel_sword_puzzle]
       - [0x3AEF0, c, e_stairway]
       - [0x3B548, c, e_sky_entities]
