+++ config/splat.us.stnp3.yaml
@@ -46,7 +46,7 @@ segments:
       - [0xE78, .data, entrance_stage_entities]
       - [0xF0C, .data, e_castle_door]
       - [0x1008, .data, e_background_bushes_trees]
-      - [0x1104, .data, entrance_stage_entities_3]
+      - [0x1104, .data, e_trapdoor]
       - [0x1120, .data, e_jewel_sword_puzzle]
       - [0x11CC, .data, e_sky_entities]
       - [0x122C, data]
@@ -171,7 +171,7 @@ segments:
       - [0x339CC, c, e_transparent_water]
       - [0x33D24, c, e_cavern_door]
       - [0x344B4, c, entrance_weights]
-      - [0x34BE8, c, entrance_stage_entities_3]
+      - [0x34BE8, c, e_trapdoor]
       - [0x34D60, c, e_jewel_sword_puzzle]
       - [0x35790, c, e_stairway]
       - [0x35DE8, c, e_sky_entities]
