+++ src/st/e_trapdoor.h
