+++ src/st/np3/e_trapdoor.c
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "np3.h"
 
-#include "../entrance_stage_entities_3.h"
+#include "../e_trapdoor.h"
